// frontend/types/sharingOptions.ts
export interface SharingOptionsConfig {
  createSubtitles: boolean;
  generateSOP: boolean;
  generateStepByStepGuide: boolean;
  generateChapters: boolean;
  autoGenerateKeywords: boolean;
  translateVideo: boolean;
  contentBranding: string;
  publishAPC: boolean;
  navigateToReviewPage: boolean;
}